fix: validate supplier relation with another product

+++ product-api/src/main/java/br/com/gabriel/product/core/services/supplier/DeleteSupplierByIdService.java
@@ -4,6 +4,7 @@ import br.com.gabriel.product.application.rest.request.EntityIdRequest;
 import br.com.gabriel.product.application.rest.response.EmptyResponse;
 import br.com.gabriel.product.core.domain.ValidationException;
 import br.com.gabriel.product.core.services.ExecutableService;
+import br.com.gabriel.product.infra.db.repositories.JpaProductRepository;
 import br.com.gabriel.product.infra.db.repositories.JpaSupplierRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,12 +16,17 @@ import java.util.Objects;
 public class DeleteSupplierByIdService implements ExecutableService<EntityIdRequest, EmptyResponse> {
 
   private final JpaSupplierRepository repository;
+  private final JpaProductRepository productRepository;
 
 
   @Override public EmptyResponse execute(final EntityIdRequest input) {
 
     if(Objects.isNull(input.id())) {
       throw new ValidationException("The id must be informed");
+    }
+
+    if(this.productRepository.existsBySupplierId(input.id())) {
+      throw new ValidationException("You cannot delete this supplier because it's already defined by a product.");
     }
 
     this.repository.deleteById(input.id());
